Fix route 53 creation for the ALB
For some reason the output of ansible changes the data structure thus old one
no longer be able to create the DNS entry. This is working with ansible 2.3.

This is working with ansible from 2.2.2.0 up to 2.3.1.0

Also update the payment gwy vault dummy on sandbox to allow the drupal build
(continued from previous patch)