Skip creating private route 53 for elb
This old code is not extendable and left from dpe era. Most of the time there
is no need to create route 53 entry for a ALB and used in that role. This
should be defered to the real code in playbook that actually uses it by getting
facts of elb and create it on demand.

Why - because thinking in the `service` term, one ELB can be used by many
services which has different service endpoints. Thus the hard-coded entry here
helps nobody and very little use.